+++ R/subset_draws.R
@@ -159,6 +159,15 @@ subset_draws.draws_rvars <- function(x, variable = NULL, iteration = NULL,
 
 #' @rdname subset_draws
 #' @export
+subset_draws.rvar <- function(x, variable = NULL, ...) {
+  if (!is.null(variable)) {
+    stop_no_call("Cannot subset an rvar by variable.")
+  }
+  subset_draws(draws_rvars(x = x), ...)$x
+}
+
+#' @rdname subset_draws
+#' @export
 subset.draws <- function(x, ...) {
   subset_draws(x, ...)
 }