+++ tests/testthat/test-thin_draws.R
@@ -5,3 +5,9 @@ test_that("thin_draws works correctly", {
   expect_error(thin_draws(x, -1), "'thin' must be a positive integer")
   expect_error(thin_draws(x, 1000), "'thin' must be smaller than")
 })
+
+test_that("thin_draws works on rvars", {
+  x <- example_draws()
+
+  expect_equal(thin_draws(as_draws_rvars(x)$theta, 10L), as_draws_rvars(thin_draws(x, 10L))$theta)
+})