+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### Enhancements
 
+* Implement `subset_draws()`, `thin_draws()`, and `resample_draws()` for `rvar`s (#225).
+* Allow `weights` to be optional in `resample_draws()` (#225).
 * Add an implementation of `drop()` for `rvar`s.
 
 ### Bug Fixes